+++ .changeset/ten-toes-clean.md
@@ -1,0 +1,5 @@
+---
+'@leafygreen-ui/pipeline': patch
+---
+
+Reorganizes directory structure internally. No external changes.
