+++ packages/pipeline/src/PipelineContext/index.tsx
@@ -2,7 +2,7 @@ import { createContext } from 'react';
 
 import { Theme } from '@leafygreen-ui/lib';
 
-import { Size } from './types';
+import { Size } from '../types';
 
 interface PipelineData {
   theme: Theme;
