+++ packages/pipeline/src/Stage/index.ts
@@ -1,0 +1,1 @@
+export { Stage } from './Stage';
