+++ packages/pipeline/src/TooltipText/TooltipText.styles.ts
@@ -1,0 +1,12 @@
+import { css } from '@leafygreen-ui/emotion';
+
+export const tooltipTextStyles = css`
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  gap: 2px;
+
+  span {
+    display: inline-flex;
+  }
+`;
