+++ packages/pipeline/src/TooltipText/index.ts
@@ -1,0 +1,1 @@
+export { TooltipText } from './TooltipText';
