+++ packages/pipeline/src/index.ts
@@ -1,3 +1,3 @@
-export { default as Pipeline } from './Pipeline';
-export { default as Stage } from './Stage';
+export { default as Pipeline } from './Pipeline/Pipeline';
+export { Stage } from './Stage';
 export { type PipelineProps, Size } from './types';
